add register,events; modify display header,footer

+++ points/apps/course/views.py
@@ -8,6 +8,13 @@ from uliweb.orm import get_model
 from forms import C_Deps_PForm as CDPForm
 from forms import C_Deps_CForm as CDCForm
 from forms import CoursesForm
+from points.models import events
+
+from uliweb import function
+require_login = function('require_login')
+from uliweb.contrib.auth.views  import login
+
+
 
 @expose('/course/')
 def index_c():
@@ -16,6 +23,8 @@ def index_c():
 
 @expose('/course/add_c')
 def add_c():
+        if require_login():
+             return redirect(url_for(login))
         form = CoursesForm()
         if request.method == 'GET':
             return {'form':form}
@@ -27,6 +36,11 @@ def add_c():
                 if c:
                    return '<a href=/course/>添加错误,重名</a>'
                 n.save()
+                ne = events()
+                ne.username = request.user
+                ne.action = '增加了课程'
+                ne.objs = form.data.c_name
+                ne.save()
                 return '<a href="/">添加完成</a>'
             else:
                 message='错误'
@@ -34,17 +48,24 @@ def add_c():
  
 @expose('/course/edit_c/<c_name>/<id>')
 def edit_c(c_name,id):
-	if request.method == 'GET': 
+    if require_login():
+          return redirect(url_for(login))
+    if request.method == 'GET': 
 		c = mcourses.get(mcourses.c.id == id)
 		form = CoursesForm(data ={'c_name':c.c_name,'c_desc':c.c_desc})
 		return {'form':form}	
-	elif request.method == 'POST':
+    elif request.method == 'POST':
             form = CoursesForm()
             flag = form.validate(request.params)
             if flag:
                 n=mcourses.get(int(id))
                 n.c_desc = form.data.c_desc
                 n.save()
+                ne = events()
+                ne.username = request.user
+                ne.action = '修改了课程'
+                ne.objs = form.data.c_name
+                ne.save()
                 return '<a href="/">添加完成</a>'
             else:
                 message='错误'
@@ -62,12 +83,21 @@ def display_c(c_name,id):
 	
 @expose('/course/delete_c/<id>')
 def delete_c(id):
-	c=mcourses.get(int(id))
-	c.delete()
-	return '<a href="/">删除完成</a>'
+    if require_login():
+        return redirect(url_for(login))
+    c=mcourses.get(int(id))
+    c.delete()
+    ne = events()
+    ne.username = request.user
+    ne.action = '删除了课程'
+    ne.objs = c.c_name
+    ne.save()
+    return '<a href="/">删除完成</a>'
 ######################################
 @expose('/course/add_cc/<c_name>')
 def add_cc(c_name):
+    	if require_login():
+        	return redirect(url_for(login))
         form = CDCForm()
         if request.method == 'GET':
             return {'form':form,'c_name':c_name}
@@ -78,12 +108,19 @@ def add_cc(c_name):
                 n.c_name = c_name
                 n.save()
                 return '<a href="/">添加完成</a>'
+                ne = events()
+                ne.username = request.user
+                ne.action = '增加了课程依赖'
+                ne.objs = c_name
+                ne.save()
             else:
                 message='错误'
                 return {'form':form}
 
 @expose('/course/add_cp/<c_name>')
 def add_cp(c_name):
+    	if require_login():
+        	return redirect(url_for(login))
         form = CDPForm()
         if request.method == 'GET':
             return {'form':form,'c_name':c_name}
@@ -93,6 +130,11 @@ def add_cp(c_name):
                 n = cdp(**form.data)
                 n.c_name = c_name
                 n.save()
+                ne = events()
+                ne.username = request.user
+                ne.action = '增加了知识点依赖'
+                ne.objs = c_name
+                ne.save()
                 return '<a href="/">添加完成</a>'
             else:
                 message='错误'
