+++ points/apps/points/views.py
@@ -3,13 +3,17 @@ from uliweb import expose
 from uliweb import settings
 from models import mpoints
 from models import mdeps
+from models import events
 from uliweb.orm import get_model
 from forms import DepsForm
+from uliweb import function
+require_login = function('require_login')
+from uliweb.contrib.auth.views  import login
 
 @expose('/')
 def index():
-	main = '知识点'
-	return {'main':main}
+	mains = events.all().order_by(events.c.datetime.desc()).limit(10)
+	return {'mains':mains}
 
 @expose('/points/')
 def index_p():
@@ -18,6 +22,8 @@ def index_p():
 
 @expose('/points/add_p')
 def add_p():
+        if require_login():
+			return redirect(url_for(login))
         from forms import PointsForm
         form = PointsForm()
         if request.method == 'GET':
@@ -30,6 +36,12 @@ def add_p():
                 if p:
                     return '<a href=/points/>添加错误，重名</a>'
                 n.save()
+                ne = events()
+                ne.username = request.user
+                ne.action = '增加了知识点'
+                ne.objs = form.data.p_name
+                ne.save()
+
                 return '<a href="/points/">添加完成</a>'
             else:
                 message='错误'
@@ -37,6 +49,8 @@ def add_p():
  
 @expose('/points/edit_p/<p_name>/<id>')
 def edit_p(p_name,id):
+	if require_login():
+		return redirect(url_for(login))
 	from forms import PointsForm
 	if request.method == 'GET': 
 		p = mpoints.get(mpoints.c.id == id)
@@ -51,6 +65,11 @@ def edit_p(p_name,id):
                 n.p_desc= form.data.p_desc
                 n.p_av_addr= form.data.p_av_addr
                 n.save()
+                ne = events()
+                ne.username = request.user
+                ne.action = '修改了知识点'
+                ne.objs = form.data.p_name
+                ne.save()
                 return '<a href="/">添加完成</a>'
             else:
                 message='错误'
@@ -67,12 +86,21 @@ def display_p(p_name,id):
 	
 @expose('/points/delete_p/<id>')
 def delete_p(id):
+	if require_login():
+		return redirect(url_for(login))
 	p=mpoints.get(int(id))
 	p.delete()
+	ne = events()
+	ne.username = request.user
+	ne.action = '删除了知识点'
+	ne.objs = p.p_name
+	ne.save()
 	return '<a href="/">ok</a>'
 ######################################
 @expose('/points/add_d/<p_name>')
 def add_d(p_name):
+        if require_login():
+			return redirect(url_for(login))
         form = DepsForm()
         if request.method == 'GET':
             return {'form':form,'p_name':p_name}
@@ -82,6 +110,11 @@ def add_d(p_name):
                 n = mdeps(**form.data)
                 n.d_name = p_name
                 n.save()
+                ne = events()
+                ne.username = request.user
+                ne.action = '增加了知识点依赖'
+                ne.objs = p_name
+                ne.save()
                 return '<a href="/">添加完成</a>'
             else:
                 message='错误'
